Bump minor version and add docs
