+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-saxo",
-  "version": "1.4.4",
+  "version": "1.5.0",
   "author": "Saxo Bank",
   "description": "ESLint rules created for Saxo Bank",
   "main": "src/index.js",
